Integrate unique model template regressions into existing tests
